USART.H corrects baud rate problems with P1C18F25K80 and maybe some others.
The problem was that the BRGH1 bit was not being set because the affected chips have no TX1STA register.  They instead have a TXSTA1 register and no alias register name in their respective dat files.  This caused the compiler to create an ASM that attempted to set a register that did not exist. Thus the actual register was not set